html/browsers/browsing-the-web/navigating-across-documents/006.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [6, 11, 12]
  - id: browsing-the-web.html#javascript-protocol
  - id: form-control-infrastructure.html#concept-form-submit
  - id: form-control-infrastructure.html#submit-get-action
  - id: form-control-infrastructure.html#plan-to-navigate
  - id: form-control-infrastructure.html#planned-navigation

description: |
  This test is operationally equivalent to 005.html with the exeption that:
    - This test creates a form element with a `target` of `test` and an `action` of `click.html`
    - the onclick handler of the anchor element in 006.html selects the form and submits that in the DOM and submits it with the click.html action on the test iframe target instead of updating the window.location of the iframe directly.
    - just like 005.html, this tests checks to make sure the href is handled and the onclick's behaviour is aborted before reaching maturity per step 6 of 7.8.1

/-->
<title>Link with onclick form submit is aborted by href navigation </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe name="test"></iframe>
<form target="test" action="click.html"></form>
<a target="test" onclick="document.forms[0].submit()" href="href.html">Test</a>
<script>
var test = async_test();

test.step(function() {
  document.querySelector("a").click();
});

window.onmessage = test.step_func(function(e) {
  assert_equals(e.data, "href");
  test.done();
});
</script>
